Add files via upload
This HTML code combines PHP and JavaScript to create a component graph using data from a MySQL database. Here is a summary of the code:

Connect to the database:
The "connection.php" file obviously contains the connection settings to the MySQL database.

Execute SQL query:
A SQL query is used to extract data from the "vulnerable" table. The results are stored in a PHP table.

Display data in an HTML table:
The results of the SQL query are displayed in an HTML table. Also, the values are stored in a PHP array for later use in the chart.

Convert PHP data to JavaScript:
PHP data is converted to a JavaScript array using the json_encode function.

Create a chart using Google Charts:
Using Google Charts, a Column Chart is created. The JavaScript array with the data is used to supply the chart data.

Display Chart:
Chart is displayed in <div> element with id "chart_div". The chart contains two columns, one for "Single_Parent_Family" and one for "Elderly".